Create a new user

+++ src/components/movies/ReactMovieTable/index.tsx
@@ -1,18 +1,19 @@
 import type { ReactTableProps } from '@components/shared/ReactTable'
 import type { MovieModel } from '@models'
-import type { InputEventHandler } from '@ts/misc'
+import type { InputEventHandler } from '@ts/types'
 
 import { ReactTable } from '@components/shared/ReactTable'
 import { useStore } from '@nanostores/react'
+import { addMessageToContext } from '@store/message'
 import {
   $contextMovieList,
   deleteMovieOnListContext,
   setMovieListOnContext,
   updateSelectedMovieOnContext
 } from '@store/movie'
-import { API_METHODS, API_URL } from '@ts/constants'
-import { parseModelToFormData } from '@ts/parsers'
-import { Button, Input } from 'antd'
+import { API_METHODS, API_URL, HTTP_STATUS } from '@ts/constants'
+import { parseModelToFormData, parseResponseErrorToMessage } from '@ts/parsers'
+import { Button, Input, Typography } from 'antd'
 import { type FC, useEffect, useMemo, useState } from 'react'
 
 export const ReactMovieTable: FC<ReactTableProps<MovieModel>> = ({ columns, dataSource }) => {
@@ -40,25 +41,38 @@ export const ReactMovieTable: FC<ReactTableProps<MovieModel>> = ({ columns, data
   useEffect(() => setMovieListOnContext(dataSource ?? []), [dataSource])
 
   const handleSearch: InputEventHandler = searchEvent => setSearchParam(searchEvent.target.value)
+  const memoizedSeach = useMemo(
+    () => (movieListInContext.length > 0 ? <Input onChange={handleSearch} /> : null),
+    [movieListInContext]
+  )
 
   const handleEdit = (_movieToEdit: MovieModel) => updateSelectedMovieOnContext(_movieToEdit)
 
   const handleDelete = async (id: string) => {
     const movieIdToDelete = parseModelToFormData({ id })
 
-    await fetch(API_URL.MOVIES, {
+    const movieDeleteResponse = await fetch(API_URL.MOVIES, {
       body: movieIdToDelete,
       method: API_METHODS.DELETE
     })
 
-    deleteMovieOnListContext(id)
-    updateSelectedMovieOnContext(null)
+    if (movieDeleteResponse.status !== HTTP_STATUS.OK) {
+      const errorMessage = await parseResponseErrorToMessage(movieDeleteResponse)
+      addMessageToContext({ content: errorMessage, type: 'error' })
+    } else {
+      deleteMovieOnListContext(id)
+      updateSelectedMovieOnContext(null)
+      addMessageToContext({ content: 'Movie deleted', type: 'success' })
+    }
   }
 
   return (
-    <>
-      <Input onChange={handleSearch} />
+    <section>
+      <Typography.Title level={2} style={{ textAlign: 'center' }}>
+        List of movies
+      </Typography.Title>
+      {memoizedSeach}
       {memoizedTable}
-    </>
+    </section>
   )
 }
